+++ src/components/common/DatePicker/DatePicker.module.css
@@ -28,8 +28,3 @@
   flex-shrink: 0;
   margin-right: 0.5rem;
 }
-
-.wave {
-  display: block;
-  margin: auto 0;
-}
