+++ src/components/common/DatePicker/DatePickerRHF.tsx
@@ -1,34 +1,6 @@
-import React, { forwardRef, useState } from "react";
+import React, { useState } from "react";
 import { Controller, useFormContext } from "react-hook-form";
-import "react-datepicker/dist/react-datepicker.css";
-import "./DatePicker.css";
-import styles from "./DatePicker.module.css";
-import CalendarIcon from "@/assets/icon-calendar.svg?react";
-
-const DatePicker = React.lazy(async () => {
-  const [{ default: ReactDatePicker }, { default: ko }] = await Promise.all([import("react-datepicker"), import("date-fns/locale/ko")]);
-
-  const CustomInput = forwardRef((props, ref: React.ForwardedRef<HTMLInputElement>) => {
-    return (
-      <div className={styles["calendar-input-wrap"]}>
-        <CalendarIcon />
-        <input {...props} ref={ref} type="text" className={styles["calendar-input"]} />
-      </div>
-    );
-  });
-
-  return {
-    default: (props) => (
-      <ReactDatePicker
-        locale={ko}
-        customInput={<CustomInput />}
-        autoComplete="off"
-        isClearable
-        {...props}
-      />
-    ),
-  };
-});
+import DatePicker from "./DatePicker";
 
 interface PropsType {
   type: "period" | "dateWithTime";
@@ -87,7 +59,7 @@ const DatePickerRHF: React.FC<PropsType> = ({ type }) => {
               )}
               />
           </div>
-          <span className={styles["wave"]}>~</span>
+          <span style={{display: "block", margin: "auto 0"}}>~</span>
           <div>
             <Controller
               control={control}
